requirements.txt
stravalib
plotly
pandas
streamlit
python-dotenv
polars
openai
requests
tavily-python
kaleido
